+++ app/src/main/java/com/limjihoon/myhero/fragment/HomeFragment.kt
@@ -1,65 +1,140 @@
-package com.limjihoon.myhero.adapter
+package com.limjihoon.myhero.fragment
 
-import android.content.Context
+import android.app.AlertDialog
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
+import android.widget.Button
+import android.widget.EditText
 import android.widget.Toast
-import androidx.recyclerview.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import com.limjihoon.myhero.R
+import com.limjihoon.myhero.activitis.ChatBotActivity
+import com.limjihoon.myhero.activitis.MapActivity
+import com.limjihoon.myhero.adapter.TodoRecyclerAdapter
+import com.limjihoon.myhero.data.Member2
 import com.limjihoon.myhero.data.Todo
-import com.limjihoon.myhero.databinding.RecyclHomeBinding
+import com.limjihoon.myhero.databinding.FragmentHomeBinding
 import com.limjihoon.myhero.network.RetrofitHelper
 import com.limjihoon.myhero.network.RetrofitService
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class TodoRecyclerAdapter(
-    private val context: Context,
-    private val items: MutableList<Todo>
-) : RecyclerView.Adapter<TodoRecyclerAdapter.VH>() {
+class HomeFragment : Fragment() {
+    lateinit var binding: FragmentHomeBinding
+    private var uid = ""
+    var items = mutableListOf<Todo>()
 
-    inner class VH(val binding: RecyclHomeBinding) : RecyclerView.ViewHolder(binding.root) {
-        init {
-            binding.btnDelete.setOnClickListener {
-                val position = adapterPosition
-                if (position != RecyclerView.NO_POSITION) {
-                    val item = items[position]
-                    deleteTodoItem(item.no, position)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentHomeBinding.inflate(layoutInflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        binding.fabtn.setOnClickListener { startActivity(Intent(requireContext(), ChatBotActivity::class.java)) }
+        binding.creatTodo.setOnClickListener { listCreate() }
+        binding.createMap.setOnClickListener { startActivity(Intent(requireContext(), MapActivity::class.java)) }
+
+        // items를 어댑터에 전달하여 초기화합니다.
+        binding.recy.adapter = TodoRecyclerAdapter(requireContext(), items)
+    }
+
+    override fun onResume() {
+        super.onResume()
+        fetchData()
+    }
+
+    private fun fetchData() {
+        val retrofit = RetrofitHelper.getRetrofitInstance("http://myhero.dothome.co.kr")
+        val retrofitService = retrofit.create(RetrofitService::class.java)
+
+        retrofitService.getMember("asdfghjklzzxcvbnmqwer1").enqueue(object : Callback<Member2> {
+            override fun onResponse(call: Call<Member2>, response: Response<Member2>) {
+                val data = response.body()
+                data?.let {
+                    binding.nickname.text = it.nickname
+                    binding.level.text = "Lv : ${it.level}"
+                    binding.tvExp.text = "${it.exp}/50"
+                    binding.coin.text = "${it.coin} COIN"
+                    uid = it.uid
+
+                    if (it.hero == 1) {
+                        binding.hero.setImageResource(R.drawable.level_up_char1)
+                    }
+
+                    fetchTodos()
                 }
             }
-        }
 
-        private fun deleteTodoItem(no: Int, position: Int) {
+            override fun onFailure(call: Call<Member2>, t: Throwable) {
+                Log.d("error", "${t.message}")
+            }
+        })
+    }
+
+    private fun fetchTodos() {
+        val retrofit = RetrofitHelper.getRetrofitInstance("http://myhero.dothome.co.kr")
+        val retrofitService = retrofit.create(RetrofitService::class.java)
+
+        retrofitService.getTodo(uid).enqueue(object : Callback<List<Todo>> {
+            override fun onResponse(call: Call<List<Todo>>, response: Response<List<Todo>>) {
+                val data = response.body()
+                data?.let {
+                    items.clear()
+                    items.addAll(it)
+                    binding.recy.adapter?.notifyDataSetChanged()
+                }
+            }
+
+            override fun onFailure(call: Call<List<Todo>>, t: Throwable) {
+                Log.d("etodo", "${t.message}")
+            }
+        })
+    }
+
+    private fun listCreate() {
+        val builder = AlertDialog.Builder(requireContext())
+        val inflater = layoutInflater
+        val dialogView = inflater.inflate(R.layout.custum_dialog_input_todo, null)
+        builder.setView(dialogView)
+
+        val dialog = builder.create()
+        dialog.show()
+
+        val dialogButton: Button = dialogView.findViewById(R.id.confirmButton)
+        val dialogButton2: Button = dialogView.findViewById(R.id.cancelButton)
+        val dialogTv = dialogView.findViewById<EditText>(R.id.scheduleEditText)
+
+        dialogButton.setOnClickListener {
+            val todo = Todo(uid, dialogTv.text.toString(), 0, 0)
+
             val retrofit = RetrofitHelper.getRetrofitInstance("http://myhero.dothome.co.kr")
             val retrofitService = retrofit.create(RetrofitService::class.java)
 
-            retrofitService.deleteTodo(no).enqueue(object : Callback<String> {
+            retrofitService.insertTodo(todo).enqueue(object : Callback<String> {
                 override fun onResponse(call: Call<String>, response: Response<String>) {
-                    if (response.isSuccessful && response.body() == "삭제 성공") {
-                        items.removeAt(position)
-                        notifyItemRemoved(position)
-                        Toast.makeText(context, "삭제 성공", Toast.LENGTH_SHORT).show()
-                    } else {
-                        Toast.makeText(context, "삭제 실패", Toast.LENGTH_SHORT).show()
-                    }
+                    val data = response.body()
+                    Toast.makeText(requireContext(), "$data", Toast.LENGTH_SHORT).show()
+                    dialog.dismiss()
+                    fetchTodos() // 새 할 일을 추가한 후 목록을 갱신합니다.
                 }
 
                 override fun onFailure(call: Call<String>, t: Throwable) {
-                    Toast.makeText(context, "삭제 에러: ${t.message}", Toast.LENGTH_SHORT).show()
+                    Log.d("error", "${t.message}")
                 }
             })
         }
-    }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): VH {
-        val binding = RecyclHomeBinding.inflate(LayoutInflater.from(context), parent, false)
-        return VH(binding)
-    }
-
-    override fun getItemCount(): Int = items.size
-
-    override fun onBindViewHolder(holder: VH, position: Int) {
-        val item = items[position]
-        holder.binding.tvWorkTodo.text = item.workTodo
+        dialogButton2.setOnClickListener {
+            dialog.dismiss()
+        }
     }
 }
